LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-fr-expected.txt

FAIL navigator.language should be fr. Was fr-fr.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
